requirements.txt
Flask==2.3.3
gunicorn==21.2.0
supabase
python-dotenv
